@@ -1,7 +1,7 @@
 {
   "name": "@rjdellecese/elm-ts",
   "version": "0.8.1",
-  "description": "A porting of TEA to TypeScript featuring fp-ts, rxjs6 and React",
+  "description": "A porting of TEA to TypeScript featuring fp-ts, rxjs and React",
   "files": [
     "lib",
     "es6"
@@ -52,58 +52,34 @@
     "rxjs": "^6.5.2"
   },
   "devDependencies": {
-    "@types/blessed": "^0.1.17",
-    "@types/glob": "^7.1.2",
-    "@types/history": "^4.6.2",
-    "@types/jest": "^26.0.3",
-    "@types/node": "^10.17.9",
+    "@types/glob": "^8.0.0",
+    "@types/history": "^4.7.11",
+    "@types/jest": "^29.2.4",
+    "@types/node": "^18.11.13",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
-    "@types/sinon": "^9.0.4",
-    "@typescript-eslint/eslint-plugin": "^5.44.0",
-    "@typescript-eslint/parser": "^5.44.0",
-    "blessed": "^0.1.81",
-    "chalk": "^4.1.0",
-    "docs-ts": "^0.5.1",
-    "eslint": "^8.28.0",
+    "@types/sinon": "^10.0.13",
+    "@typescript-eslint/eslint-plugin": "^5.46.1",
+    "@typescript-eslint/parser": "^5.46.1",
+    "chalk": "^4.1.2",
+    "docs-ts": "^0.7.0",
+    "eslint": "^8.29.0",
     "eslint-plugin-import": "^2.26.0",
-    "fp-ts": "^2.0.2",
-    "glob": "^7.1.6",
+    "fp-ts": "^2.13.1",
+    "glob": "^8.0.3",
     "husky": "^8.0.2",
-    "io-ts": "^2.0.0",
-    "jest": "^26.1.0",
-    "monocle-ts": "^2.0.0",
-    "prettier": "^2.0.5",
-    "pretty-quick": "^2.0.1",
+    "io-ts": "^2.2.20",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "monocle-ts": "^2.3.13",
+    "prettier": "^2.8.1",
+    "pretty-quick": "^3.1.3",
     "react": "^18.2.0",
-    "react-blessed": "^0.6.2",
     "react-dom": "^18.2.0",
-    "rxjs": "^6.5.2",
-    "sinon": "^9.0.2",
-    "ts-jest": "^26.1.1",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.5.3"
-  },
-  "jest": {
-    "preset": "ts-jest",
-    "globals": {
-      "ts-jest": {
-        "diagnostics": true
-      }
-    },
-    "bail": true,
-    "collectCoverage": true,
-    "coveragePathIgnorePatterns": [
-      "<rootDir>/test/",
-      "<rootDir>/node_modules/"
-    ],
-    "coverageReporters": [
-      "text"
-    ],
-    "roots": [
-      "<rootDir>/test/"
-    ],
-    "testMatch": null,
-    "testRegex": "(\\.|/)(test|spec)\\.tsx?$"
+    "rxjs": "^7.6.0",
+    "sinon": "^15.0.0",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
   }
 }
