@@ -1,4 +1,4 @@
-import { render } from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { programWithDebugger } from '../src/Debug/Html'
 import * as React from '../src/React'
 import * as component from './Counter'
@@ -7,4 +7,10 @@ const program = process.env.NODE_ENV === 'production' ? React.program : programW
 
 const main = program(component.init, component.update, component.view)
 
-React.run(main, dom => render(dom, document.getElementById('app')))
+const element = document.getElementById('app')
+
+if (element) {
+  React.run(main, dom => createRoot(element).render(dom))
+} else {
+  throw 'Failed to find app element'
+}
