@@ -1,6 +1,5 @@
 import { mapLeft, taskify } from 'fp-ts/lib/TaskEither'
-import { flow } from 'fp-ts/lib/function'
-import { pipe } from 'fp-ts/lib/pipeable'
+import { pipe, flow } from 'fp-ts/lib/function'
 import * as fs from 'fs'
 import Glob from 'glob'
 import { Eff } from './program'
@@ -11,7 +10,7 @@ export interface FileSystem {
   readonly glob: (pattern: string) => Eff<string[]>
 }
 
-const readFileTE = taskify<fs.PathLike, string, NodeJS.ErrnoException, string>(fs.readFile)
+const readFileTE = taskify<fs.PathLike, 'utf8', NodeJS.ErrnoException, string>(fs.readFile)
 const writeFileTE = taskify<fs.PathLike, string, NodeJS.ErrnoException, void>(fs.writeFile)
 const globTE = taskify<string, Error, string[]>(Glob)
 const toError = (e: Error): string => e.message
