@@ -6,7 +6,7 @@ describe('Sub', () => {
   it('map() should transform a Sub<A> into Sub<B>', done => {
     const subA = of('a')
 
-    return map(a => a + 'b')(subA).subscribe(v => {
+    map(a => a + 'b')(subA).subscribe(v => {
       assert.strictEqual(v, 'ab')
 
       done()
@@ -17,7 +17,7 @@ describe('Sub', () => {
     const log: string[] = []
     const subs = [of('a'), of('b'), of('c')]
 
-    return batch(subs).subscribe({
+    batch(subs).subscribe({
       next: v => log.push(v),
 
       complete: () => {
